React Practice - YTSearch App
Created a React based app as part of a React course to search YouTube
using the YouTube API and return the list of videos for a user to click
on and watch. A main video will show up on the left to actually watch
the video and can be updated as videos from the list are selected.
Typing in the search bar will trigger off new YouTube API searches. Did
not commit Node Modules with this initial commit of my work as it was an
extremely high number of files.